Created new tests for F2 Primitives
