+++ geminidr/f2/test/conftest.py
@@ -1,0 +1,17 @@
+
+import pytest
+import os
+
+
+@pytest.fixture
+def test_path():
+
+    try:
+        path = os.environ['TEST_PATH']
+    except KeyError:
+        pytest.skip("Could not find environment variable: $TEST_PATH")
+
+    if not os.path.exists(path):
+        pytest.skip("Could not find path stored in $TEST_PATH: {}".format(path))
+
+    return path
